+++ lib/src/api/weather_api_client.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'package:flutter_weather/src/model/Weather.dart';
+import 'package:flutter_weather/src/model/weather.dart';
 import 'package:meta/meta.dart';
 import 'package:http/http.dart' as http;
