+++ lib/src/bloc/weather_bloc.dart
@@ -1,10 +1,10 @@
 import 'package:bloc/bloc.dart';
-import 'package:flutter_weather/src/model/Weather.dart';
+import 'package:flutter_weather/src/model/weather.dart';
 import 'package:meta/meta.dart';
 
 import 'package:flutter_weather/src/bloc/weather_event.dart';
 import 'package:flutter_weather/src/bloc/weather_state.dart';
-import 'package:flutter_weather/src/repository/WeatherRepository.dart';
+import 'package:flutter_weather/src/repository/weather_repository.dart';
 
 class WeatherBloc extends Bloc<WeatherEvent, WeatherState> {
   final WeatherRepository weatherRepository;
