+++ lib/src/bloc/weather_state.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_weather/src/model/Weather.dart';
+import 'package:flutter_weather/src/model/weather.dart';
 import 'package:meta/meta.dart';
 import 'package:equatable/equatable.dart';
