+++ lib/src/model/weather.dart
