+++ lib/src/screens/weather_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_weather/src/api/WeatherApiClient.dart';
+import 'package:flutter_weather/src/api/weather_api_client.dart';
 import 'package:flutter_weather/src/bloc/weather_bloc.dart';
 import 'package:flutter_weather/src/bloc/weather_event.dart';
 import 'package:flutter_weather/src/bloc/weather_state.dart';
-import 'package:flutter_weather/src/repository/WeatherRepository.dart';
-import 'package:flutter_weather/src/api/ApiKey.dart';
+import 'package:flutter_weather/src/repository/weather_repository.dart';
+import 'package:flutter_weather/src/api/api_keys.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_weather/src/widgets/weather_widget.dart';
 import 'package:http/http.dart' as http;
